regmap: debugfs: Optimize seeking within blocks of registers
Optimize this so that we can better guess where to start scanning
from.  We know the length of the register field format, therefore
given the file pointer position align to the nearest register
field and scan from there onwards.  We round down in this calculation
and we let the rest of the code figure out where to start scanning
from.

Change-Id: I2515e0981ef6b17f784dd0e389d0d4a13d47123c
Signed-off-by: Dimitris Papastamos <dp@opensource.wolfsonmicro.com>