switch: arizona: Suppress duplicate JACKDET reports
In cases where we see a brief (dis)connection of the jack detection signals
we may see a noop jack insertion or removal where the jack has returned to
the original state by the time the interrupt is serviced. Suppress these
events in order to save work and avoid confusing the rest of the code.

Change-Id: I496a72c239988e8b0939f9c86a216fa85cc427c3
Signed-off-by: Mark Brown <broonie@opensource.wolfsonmicro.com>