arm/dt: falcon: Configure PMIC Resets
This changes the DTS so that the kernel driver will
configure the PMIC resets just as the Bootloader
does allowing for the driver to handle the Stage 1
Interrupt and perform a proper shutdown if things
are running fine at the time.  It setups the Power
Key Trigger only and disables the Volume Down Trigger.

Change-Id: I8d8e38e18cb0a475b06f2699134d7537596547d5
Signed-off-by: Kenny Kessler <kenny.kessler@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/548699
SLT-Approved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Submit-Approved: Jira Key <jirakey@motorola.com>