power: qpnp-bms: Quick SOC Calculation
There is a Race at Probe where part of the Battery
Metering Driver can accidentally use Uninitialized
values. When that happens it correctly bounds it
to 0% but then thinks that is the last reported
value and continues to use it.

To fix this run the calculate_soc_work sooner at
probe time.  Also return 100% if we try and report
an SOC before Initialization of hte parameters occurs

Change-Id: Idc25f96d446e0ac02928fe307c2f3a95f883354a
Signed-off-by: Kenny Kessler <kenny.kessler@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/564910
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Adam Ruehl <aruehl@motorola.com>
Reviewed-by: Christopher Fries <c.fries@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>