ASoC TPA6165: fix inconsistent button press det
button press state needs to be turned off when button release
is recorded. if it is left ON when the IC is put to full sleep
button press detect state is turned OFF actually in IC,
but the driver still thinks its on and doesnt enable it back
again, so button presses are missed.

Change-Id: I5786bcd1dde84bf1ffd7d771041b5cfd371b189d
Signed-off-by: prabhu annabathula <prabhu.annabathula@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/569469
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>