arm/dts: msm8974: Enable sleep in memory retention for Victara
For the Victara display, the display spec now calls for us to enable the
sleep in memory retention feature.  This feature can not be MTP'd on, so
we need to turn it on in the init sequence.

Change-Id: I41a6850888b921cee072d5f54705806798d6bfef
Signed-off-by: Zac Crosby <azc095@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/604492
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Bang Nguyen <bangnguyen@motorola.com>
Reviewed-by: Stephen Rossbach <rossbach@motorola.com>
SLTApproved: Stephen Rossbach <rossbach@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>