video: fb_quickdraw: add buffer alignment helper functions
To prevent display corruption when drawing non-fullscreen buffers
to the display panel the buffers need to be aligned along the
x-axis per the panel "column alignment" specification. For example,
if a panel has a 8-pixel alignment restriction buffers must start
at a x-coordinate that is divisible by 8 with a width that is also
divisible by 8. Buffers are auto-corrected by shifting to the nearest
x-coordinate that is divisible by the panel's column alignment
requirement. Any buffer that has a width that is not divisible by the
panel's column alignment requirement is considered invalid.

This adds the helper functions to the generic fb_quickdraw engine
that can be used by the panel specific quickdraw implementations.

Change-Id: I45787e2f66984893fe2531b836ab89173a24b7b0
Signed-off-by: Joe Swantek <jswantek@motorola.com>
Reviewed-on: http://gerrit.mot.com/631899
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Patrick Auchter <auchter@motorola.com>
Reviewed-by: Stephen Rossbach <rossbach@motorola.com>
Reviewed-by: Bang Nguyen <bangnguyen@motorola.com>
SLTApproved: Christopher Fries <c.fries@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>