qup_i2c: Add 1-clk pulse recovery
If the transaction is corrupted in such a way that the host
and device are out-of-sync, then the device can output
it's ACK signal indefinitely, and the 9-cycle recovery method
will clear the ACK on the first cycle, and then force a new
ACK on the 9th cycle, allowing the failure to continue with
DATA driven low.To fix this issue the 1-cycle recovery method
added to recover the bus, after the 9-cycle recovery method failed.
This is a workaround for specific LM3642 HW problem and may not
work with other slave devices. To enable the workaround set
qcom,extended-recovery device tree flag to 1.

Change-Id: I936c16c925ed2bcabfbf49b2590665262235b8e0
Signed-off-by: Vladimir Karpovich <cvk011c@motorola.com>
Reviewed-on: http://gerrit.mot.com/647691
Tested-by: Jira Key <jirakey@motorola.com>
SLTApproved: Christopher Fries <cfries@motorola.com>
Reviewed-by: Igor Kovalenko <igork@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>

Conflicts:
	Documentation/devicetree/bindings/i2c/i2c-qup.txt
	arch/arm/mach-msm/include/mach/board.h