misc: stm401: correct timeout during stm401 reset
The original intention of the increased i2c retry delay of 200ms after
the stm401 IC is reset is to give an increased timeout for the first
i2c transfer after the IC is reset. Correct the placement of the
i2c retry delay increase with this change to it's originally intended
location.

Change-Id: I10e1ef96b79e8fa007e9ba31eddb482582f0c303
Signed-off-by: Joe Swantek <jswantek@motorola.com>
Reviewed-on: http://gerrit.mot.com/709619
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Robert Wiegand <Robert.P.Wiegand@motorola.com>
Reviewed-by: James Rumpler <w12004@motorola.com>
Reviewed-by: Fred Fettinger <fettinge@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>
Reviewed-by: Jared Suttles <jsuttles@motorola.com>
Reviewed-by: Stephen Rossbach <rossbach@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>
SLTApproved: Connie Zhao <czhao1@motorola.com>